fixed big errors in divide and changed types to int64

+++ include/input_query.hpp
@@ -17,30 +17,30 @@ enum class PossibleRounding : std::uint8_t {
 };
 class InputQuery {
   public:
-    InputQuery(std::int32_t cnt_for_integer, std::int32_t cnt_for_fractional,
-               std::int32_t first_number,
+    InputQuery(std::int64_t cnt_for_integer, std::int64_t cnt_for_fractional,
+               std::int64_t first_number,
                PossibleRounding cur_rounding = PossibleRounding::TOWARD_ZERO,
-               std::int32_t second_number = 0,
+               std::int64_t second_number = 0,
                PossibleOperations cur_operation = PossibleOperations::NO_OPER);
 
     [[nodiscard]]
-    std::int32_t get_first_number() const noexcept;
+    std::int64_t get_first_number() const noexcept;
     [[nodiscard]]
-    std::int32_t get_second_number() const noexcept;
+    std::int64_t get_second_number() const noexcept;
     [[nodiscard]]
-    std::int32_t get_cnt_for_integer() const noexcept;
+    std::int64_t get_cnt_for_integer() const noexcept;
     [[nodiscard]]
-    std::int32_t get_cnt_for_fractional() const noexcept;
+    std::int64_t get_cnt_for_fractional() const noexcept;
     [[nodiscard]]
     PossibleOperations get_cur_operation() const noexcept;
     [[nodiscard]]
     PossibleRounding get_cur_rounding() const noexcept;
 
   private:
-    std::int32_t _cnt_for_integer;
-    std::int32_t _cnt_for_fractional;
-    std::int32_t _first_number;
-    std::int32_t _second_number;
+    std::int64_t _cnt_for_integer;
+    std::int64_t _cnt_for_fractional;
+    std::int64_t _first_number;
+    std::int64_t _second_number;
     PossibleRounding _cur_rounding;
     PossibleOperations _cur_operation;
 };
