+++ src/input_query.cpp
@@ -1,11 +1,11 @@
 #include "input_query.hpp"
 
 [[nodiscard]]
-std::int32_t InputQuery::get_first_number() const noexcept {
+std::int64_t InputQuery::get_first_number() const noexcept {
     return _first_number;
 }
 [[nodiscard]]
-std::int32_t InputQuery::get_second_number() const noexcept {
+std::int64_t InputQuery::get_second_number() const noexcept {
     return _second_number;
 }
 [[nodiscard]]
@@ -18,18 +18,18 @@ PossibleRounding InputQuery::get_cur_rounding() const noexcept {
 }
 
 [[nodiscard]]
-std::int32_t InputQuery::get_cnt_for_integer() const noexcept {
+std::int64_t InputQuery::get_cnt_for_integer() const noexcept {
     return _cnt_for_integer;
 }
 [[nodiscard]]
-std::int32_t InputQuery::get_cnt_for_fractional() const noexcept {
+std::int64_t InputQuery::get_cnt_for_fractional() const noexcept {
     return _cnt_for_fractional;
 }
 
-InputQuery::InputQuery(std::int32_t cnt_for_integer,
-                       std::int32_t cnt_for_fractional,
-                       std::int32_t first_number, PossibleRounding cur_rounding,
-                       std::int32_t second_number,
+InputQuery::InputQuery(std::int64_t cnt_for_integer,
+                       std::int64_t cnt_for_fractional,
+                       std::int64_t first_number, PossibleRounding cur_rounding,
+                       std::int64_t second_number,
                        PossibleOperations cur_operation)
     : _cnt_for_integer(cnt_for_integer),
       _cnt_for_fractional(cnt_for_fractional), _first_number(first_number),
